HR Candidate Evaluation/requirements.txt
fastapi
uvicorn
psycopg2-binary
PyPDF2
ollama
python-multipart
